# townService/src/town/TownsController.ts
import assert from 'assert';
import {
  Body,
  Controller,
  Delete,
  Example,
  Get,
  Header,
  Patch,
  Path,
  Post,
  Response,
  Route,
  Tags,
} from 'tsoa';

import { title } from 'process';
import { Town, TownCreateParams, TownCreateResponse } from '../api/Model';
import InvalidParametersError from '../lib/InvalidParametersError';
import CoveyTownsStore from '../lib/TownsStore';
import {
  ConversationArea,
  CoveyTownSocket,
  TownSettingsUpdate,
  ViewingArea,
  PosterSessionArea,
  WatchTogetherArea,
  Video,
} from '../types/CoveyTownSocket';
import PosterSessionAreaReal from './PosterSessionArea';
import { isPosterSessionArea, isWatchTogetherArea } from '../TestUtils';
import WatchTogetherAreaModel from './WatchTogetherArea';
import getVideoDetail from '../api/YoutubeAPI';

/**
 * This is the town route
 */
@Route('towns')
@Tags('towns')
// TSOA (which we use to generate the REST API from this file) does not support default exports, so the controller can't be a default export.
// eslint-disable-next-line import/prefer-default-export
export class TownsController extends Controller {
  private _townsStore: CoveyTownsStore = CoveyTownsStore.getInstance();

  /**
   * List all towns that are set to be publicly available
   *
   * @returns list of towns
   */
  @Get()
  public async listTowns(): Promise<Town[]> {
    return this._townsStore.getTowns();
  }

  /**
   * Create a new town
   *
   * @param request The public-facing information for the new town
   * @example request {"friendlyName": "My testing town public name", "isPubliclyListed": true}
   * @returns The ID of the newly created town, and a secret password that will be needed to update or delete this town.
   */
  @Example<TownCreateResponse>({ townID: 'stringID', townUpdatePassword: 'secretPassword' })
  @Post()
  public async createTown(@Body() request: TownCreateParams): Promise<TownCreateResponse> {
    const { townID, townUpdatePassword } = await this._townsStore.createTown(
      request.friendlyName,
      request.isPubliclyListed,
      request.mapFile,
    );
    return {
      townID,
      townUpdatePassword,
    };
  }

  /**
   * Updates an existing town's settings by ID
   *
   * @param townID  town to update
   * @param townUpdatePassword  town update password, must match the password returned by createTown
   * @param requestBody The updated settings
   */
  @Patch('{townID}')
  @Response<InvalidParametersError>(400, 'Invalid password or update values specified')
  public async updateTown(
    @Path() townID: string,
    @Header('X-CoveyTown-Password') townUpdatePassword: string,
    @Body() requestBody: TownSettingsUpdate,
  ): Promise<void> {
    const success = this._townsStore.updateTown(
      townID,
      townUpdatePassword,
      requestBody.friendlyName,
      requestBody.isPubliclyListed,
    );
    if (!success) {
      throw new InvalidParametersError('Invalid password or update values specified');
    }
  }

  /**
   * Deletes a town
   * @param townID ID of the town to delete
   * @param townUpdatePassword town update password, must match the password returned by createTown
   */
  @Delete('{townID}')
  @Response<InvalidParametersError>(400, 'Invalid password or update values specified')
  public async deleteTown(
    @Path() townID: string,
    @Header('X-CoveyTown-Password') townUpdatePassword: string,
  ): Promise<void> {
    const success = this._townsStore.deleteTown(townID, townUpdatePassword);
    if (!success) {
      throw new InvalidParametersError('Invalid password or update values specified');
    }
  }

  /**
   * Creates a conversation area in a given town
   * @param townID ID of the town in which to create the new conversation area
   * @param sessionToken session token of the player making the request, must match the session token returned when the player joined the town
   * @param requestBody The new conversation area to create
   */
  @Post('{townID}/conversationArea')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async createConversationArea(
    @Path() townID: string,
    @Header('X-Session-Token') sessionToken: string,
    @Body() requestBody: ConversationArea,
  ): Promise<void> {
    const town = this._townsStore.getTownByID(townID);
    if (!town?.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid values specified');
    }
    const success = town.addConversationArea(requestBody);
    if (!success) {
      throw new InvalidParametersError('Invalid values specified');
    }
  }

  /**
   * Creates a viewing area in a given town
   *
   * @param townID ID of the town in which to create the new viewing area
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   * @param requestBody The new viewing area to create
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          viewing area could not be created
   */
  @Post('{townID}/viewingArea')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async createViewingArea(
    @Path() townID: string,
    @Header('X-Session-Token') sessionToken: string,
    @Body() requestBody: ViewingArea,
  ): Promise<void> {
    const town = this._townsStore.getTownByID(townID);
    if (!town) {
      throw new InvalidParametersError('Invalid values specified');
    }
    if (!town?.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid values specified');
    }
    const success = town.addViewingArea(requestBody);
    if (!success) {
      throw new InvalidParametersError('Invalid values specified');
    }
  }

  /**
   * Creates a poster session area in a given town
   *
   * @param townID ID of the town in which to create the new poster session area
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   * @param requestBody The new poster session area to create
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          poster session area could not be created
   */
  @Post('{townID}/posterSessionArea')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async createPosterSessionArea(
    @Path() townID: string,
    @Header('X-Session-Token') sessionToken: string,
    @Body() requestBody: PosterSessionArea,
  ): Promise<void> {
    // download file here TODO
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    // add viewing area to the town, throw error if it fails
    if (!curTown.addPosterSessionArea(requestBody)) {
      throw new InvalidParametersError('Invalid poster session area');
    }
  }

  /**
   * Gets the image contents of a given poster session area in a given town
   *
   * @param townID ID of the town in which to get the poster session area image contents
   * @param posterSessionId interactable ID of the poster session
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          poster session specified does not exist
   */
  @Patch('{townID}/{posterSessionId}/imageContents')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async getPosterAreaImageContents(
    @Path() townID: string,
    @Path() posterSessionId: string,
    @Header('X-Session-Token') sessionToken: string,
  ): Promise<string | undefined> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    const posterSessionArea = curTown.getInteractable(posterSessionId);
    if (!posterSessionArea || !isPosterSessionArea(posterSessionArea)) {
      throw new InvalidParametersError('Invalid poster session ID');
    }
    return posterSessionArea.imageContents;
  }

  /**
   * Increment the stars of a given poster session area in a given town, as long as there is
   * a poster image. Returns the new number of stars.
   *
   * @param townID ID of the town in which to get the poster session area image contents
   * @param posterSessionId interactable ID of the poster session
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          poster session specified does not exist, or if the poster session specified
   *          does not have an image
   */
  @Patch('{townID}/{posterSessionId}/incStars')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async incrementPosterAreaStars(
    @Path() townID: string,
    @Path() posterSessionId: string,
    @Header('X-Session-Token') sessionToken: string,
  ): Promise<number> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    const posterSessionArea = curTown.getInteractable(posterSessionId);
    if (!posterSessionArea || !isPosterSessionArea(posterSessionArea)) {
      throw new InvalidParametersError('Invalid poster session ID');
    }
    if (!posterSessionArea.imageContents) {
      throw new InvalidParametersError('Cant star a poster with no image');
    }
    const newStars = posterSessionArea.stars + 1;
    const updatedPosterSessionArea = {
      id: posterSessionArea.id,
      imageContents: posterSessionArea.imageContents,
      title: posterSessionArea.title,
      stars: newStars, // increment stars
    };
    (<PosterSessionAreaReal>posterSessionArea).updateModel(updatedPosterSessionArea);
    return newStars;
  }

  /**
   * Creates a watch together area in a given town
   *
   * @param townID ID of the town in which to create the new watch together area
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   * @param requestBody The new watch together area to create
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          watch together area could not be created
   */
  @Post('{townID}/watchTogetherArea')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async createWatchTogetherArea(
    @Path() townID: string,
    @Header('X-Session-Token') sessionToken: string,
    @Body() requestBody: WatchTogetherArea,
  ): Promise<void> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    // add viewing area to the town, throw error if it fails
    if (!curTown.addWatchTogetherArea(requestBody)) {
      throw new InvalidParametersError('Invalid watch together area');
    }
  }

  // /**
  //  * Gets the video information of a given watch together area in a given town
  //  *
  //  * @param townID ID of the town in which to get the watch together area video information
  //  * @param watchTogetherId interactable ID of the watch together area
  //  * @param sessionToken session token of the player making the request, must
  //  *        match the session token returned when the player joined the town
  //  *
  //  * @throws InvalidParametersError if the session token is not valid, or if the
  //  *          watch together specified does not exist
  //  */
  // @Patch('{townID}/{watchTogetherId}/videoInfo')
  // @Response<InvalidParametersError>(400, 'Invalid values specified')
  // public async getWatchTogetherVideo(
  //   @Path() townID: string,
  //   @Path() watchTogetherId: string,
  //   @Header('X-Session-Token') sessionToken: string,
  // ): Promise<Video | undefined> {
  //   const curTown = this._townsStore.getTownByID(townID);
  //   if (!curTown) {
  //     throw new InvalidParametersError('Invalid town ID');
  //   }
  //   if (!curTown.getPlayerBySessionToken(sessionToken)) {
  //     throw new InvalidParametersError('Invalid session ID');
  //   }
  //   const watchTogetherArea = curTown.getInteractable(watchTogetherId);
  //   if (!watchTogetherArea || !isWatchTogetherArea(watchTogetherArea)) {
  //     throw new InvalidParametersError('Invalid poster session ID');
  //   }
  //   return watchTogetherArea.video;
  // }

  /**
   * Gets the host ID of a given watch together area in a given town
   *
   * @param townID ID of the town in which to get the watch together area hostID
   * @param watchTogetherId interactable ID of the watch together area
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          watch together specified does not exist
   */
  @Patch('{townID}/{watchTogetherId}/hostID')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async getWatchTogetherHostID(
    @Path() townID: string,
    @Path() watchTogetherId: string,
    @Header('X-Session-Token') sessionToken: string,
  ): Promise<string | undefined> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    const watchTogetherArea = curTown.getInteractable(watchTogetherId);
    if (!watchTogetherArea || !isWatchTogetherArea(watchTogetherArea)) {
      throw new InvalidParametersError('Invalid poster session ID');
    }
    return watchTogetherArea.hostID;
  }

  // /**
  //  * Gets the playlist of a given watch together area in a given town
  //  *
  //  * @param townID ID of the town in which to get the watch together area playlist
  //  * @param watchTogetherId interactable ID of the watch together area
  //  * @param sessionToken session token of the player making the request, must
  //  *        match the session token returned when the player joined the town
  //  *
  //  * @throws InvalidParametersError if the session token is not valid, or if the
  //  *          watch together specified does not exist
  //  */
  // @Patch('{townID}/{watchTogetherId}/playList')
  // @Response<InvalidParametersError>(400, 'Invalid values specified')
  // public async getWatchTogetherPlayList(
  //   @Path() townID: string,
  //   @Path() watchTogetherId: string,
  //   @Header('X-Session-Token') sessionToken: string,
  // ): Promise<Video[]> {
  //   const curTown = this._townsStore.getTownByID(townID);
  //   if (!curTown) {
  //     throw new InvalidParametersError('Invalid town ID');
  //   }
  //   if (!curTown.getPlayerBySessionToken(sessionToken)) {
  //     throw new InvalidParametersError('Invalid session ID');
  //   }
  //   const watchTogetherArea = curTown.getInteractable(watchTogetherId);
  //   if (!watchTogetherArea || !isWatchTogetherArea(watchTogetherArea)) {
  //     throw new InvalidParametersError('Invalid poster session ID');
  //   }
  //   return watchTogetherArea.playList;
  // }

  /**
   * Push new video to the playList of the given watch together area, as long as the
   * hostID is not undefined.
   *
   * @param townID ID of the town in which to push watch together area playList
   * @param posterSessionId interactable ID of the watch together
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          poster session specified does not exist, or if the poster session specified
   *          does not have an image
   */
  @Post('{townID}/{watchTogetherId}/addVideotoPlaylist') // Fixing
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async pushWatchTogetherPlayList(
    @Path() townID: string,
    @Path() watchTogetherId: string,
    @Header('X-Session-Token') sessionToken: string,
    @Body() requestBody: { url: string },
  ): Promise<Video> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    const curPlayer = curTown.getPlayerBySessionToken(sessionToken);
    if (!curPlayer) {
      throw new InvalidParametersError('Invalid session ID');
    }
    const watchTogetherArea = curTown.getInteractable(watchTogetherId);
    if (!watchTogetherArea || !isWatchTogetherArea(watchTogetherArea)) {
      throw new InvalidParametersError('Invalid watch together ID');
    }
    if (!watchTogetherArea.hostID) {
      throw new InvalidParametersError('Cant add video to watch together with no host');
    }
    const video = await curTown.addVideoToWatchTogetherPlaylist(
      requestBody.url,
      watchTogetherArea,
      curPlayer.id,
    );
    if (!video) {
      throw new InvalidParametersError('Invalid watch together area');
    }
    return video;
  }

  /**
   * Play the next video in a given watch together area in a given town
   *
   * @param townID ID of the town in which to get the watch together area
   * @param watchTogetherId interactable ID of the watch together area
   * @param sessionToken session token of the player making the request, must
   *        match the session token returned when the player joined the town
   *
   * @throws InvalidParametersError if the session token is not valid, or if the
   *          watch together specified does not exist
   */
  @Patch('{townID}/{watchTogetherId}/playNext')
  @Response<InvalidParametersError>(400, 'Invalid values specified')
  public async watchTogetherPlayNext(
    @Path() townID: string,
    @Path() watchTogetherId: string,
    @Header('X-Session-Token') sessionToken: string,
  ): Promise<boolean> {
    const curTown = this._townsStore.getTownByID(townID);
    if (!curTown) {
      throw new InvalidParametersError('Invalid town ID');
    }
    if (!curTown.getPlayerBySessionToken(sessionToken)) {
      throw new InvalidParametersError('Invalid session ID');
    }
    const watchTogetherArea = curTown.getInteractable(watchTogetherId);
    if (!watchTogetherArea || !isWatchTogetherArea(watchTogetherArea)) {
      throw new InvalidParametersError('Invalid poster session ID');
    }
    return curTown.watchTogetherPlayNext(watchTogetherArea);
  }

  /**
   * Connects a client's socket to the requested town, or disconnects the socket if no such town exists
   *
   * @param socket A new socket connection, with the userName and townID parameters of the socket's
   * auth object configured with the desired townID to join and username to use
   *
   */
  public async joinTown(socket: CoveyTownSocket) {
    // Parse the client's requested username from the connection
    const { userName, townID } = socket.handshake.auth as { userName: string; townID: string };

    const town = this._townsStore.getTownByID(townID);
    if (!town) {
      socket.disconnect(true);
      return;
    }

    // Connect the client to the socket.io broadcast room for this town
    socket.join(town.townID);

    const newPlayer = await town.addPlayer(userName, socket);
    assert(newPlayer.videoToken);
    socket.emit('initialize', {
      userID: newPlayer.id,
      sessionToken: newPlayer.sessionToken,
      providerVideoToken: newPlayer.videoToken,
      currentPlayers: town.players.map(eachPlayer => eachPlayer.toPlayerModel()),
      friendlyName: town.friendlyName,
      isPubliclyListed: town.isPubliclyListed,
      interactables: town.interactables.map(eachInteractable => eachInteractable.toModel()),
    });
  }
}